+++ packages/unocss/package.json
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "@penxle/eslint-config": "workspace:^",
-    "tsup": "^7.1.0"
+    "tsup": "^7.1.0",
+    "vite": "^4.4.7"
   }
 }
