+++ sites/landing.penxle.com/vite.config.ts
@@ -1,8 +1,8 @@
+import { unocss } from '@penxle/unocss';
 import { svg } from '@penxle/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 import browserslist from 'browserslist';
 import { browserslistToTargets } from 'lightningcss';
-import unocss from 'unocss/vite';
 import { defineConfig } from 'vite';
 
 // eslint-disable-next-line import/no-default-export
