+++ sites/penxle.io/src/routes/+layout.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import '../styles/index.css';
+  // import 'uno.css';
   import { beforeNavigate } from '$app/navigation';
   import { updated } from '$app/stores';
   import { setupAnalytics } from '$lib/analytics';
