+++ sites/penxle.io/svelte.config.js
@@ -1,11 +1,10 @@
-import { unoPreprocess } from '@penxle/unocss';
 import vercel from '@sveltejs/adapter-vercel';
 import { vitePreprocess } from '@sveltejs/kit/vite';
 
 /** @type {import('@sveltejs/kit').Config} */
 // eslint-disable-next-line import/no-default-export
 export default {
-  preprocess: [vitePreprocess(), unoPreprocess()],
+  preprocess: vitePreprocess(),
   kit: {
     adapter: vercel({ runtime: 'edge' }),
     alias: { $assets: './src/assets' },
