+++ sites/penxle.io/vite.config.ts
@@ -1,3 +1,4 @@
+import { unocss } from '@penxle/unocss';
 import { svg } from '@penxle/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 import browserslist from 'browserslist';
@@ -19,7 +20,7 @@ export default defineConfig({
       ),
     },
   },
-  plugins: [svg(), sveltekit()],
+  plugins: [svg(), unocss(), sveltekit()],
   server: {
     host: '127.0.0.1',
     port: 4010,
